all: remove ServerRequest[T] for concrete T
Replace all occurrences of ServerRequest[*CallToolParams] and other
concrete instantiations with CallToolRequest and the like.

Make the XXXRequest types aliases, to preserve the convenience of
generics for the internal machinery (see shared.go, for example.)

TODO: either expand the aliases or unexport ServerRequest.
(The latter will be problematic for docs.)