auth: client-side OAuth flow
This is a preliminary implementation of OAuth 2.1 for the client.

It provides an http.RoundTripper, auth.HTTPTransport, that invokes
a user-provided callback of type auth.OAuthHandler.
The latter is responsible for all the OAuth work.
We will add code to make that easier in later PRs.

Much remains to be done :

Dynamic client registration is not implemented. Since it is optional,
we also need another way of supplying the client ID and secret to
this code.

Resource Indicators, as described in section 2.5.1 of the MCP spec.

And, of course, tests. We should test against fake implementations
but also, if we can find any, real reference implementations.