mcp: add OnConnectionClose.OnConnectionClose(...) callback
Adding a new callback that callers can optionally set in order to
get notified about a connection created by a `StreamableHTTPHandler`
was closed. Connections can be closed by the MCP client as part of
the regular connection lifecycle (happy path) or when there was a
connection error (e.g., a timeout).

Fixes #479