mcp: prototype implementation for SEP 1442
This PR is the result of spending a few hours going through and
attempting to implement SEP 1442
(modelcontextprotocol/modelcontextprotocol#1442).

Implemented:
 - skipping initialization
 - new protocol version and session ID treatment
 - server/discover
 - unsupported version errors
 - new _meta fields
 - client capability embedding

Still TODO:
 - client-initiated streams
 - ergonomic and documentation for handling capabilities from
   application code
 - many, many more tests (and corresponding bug fixes)

This is very much quick and dirty, though I learned a lot about the SEP
in the process. Tests pass, albeit largely because I configure the
client to keep initialization, or downgrade the protocol version.

Notably, server->client requests do work in the context of an
uninitialized session.

Additional observations will be noted in PR comments.