mcp/client: enforce JSON 2020-12 spec for schemas
https://github.com/modelcontextprotocol/go-sdk/issues/624

Both the client and the server APIs use jsonschema.Schema objects to
represent schemas. The `Schema` object only supports the JSON 2020-12
format and will fail if we can't unmarshall the schema into the JSON
2020-12 format.

This change attempts to detect when requests are using a different
schema spec, and give more descriptive errors messages when
possible,rather than the unmarshalling to jsonschema.Schema failing.