mcp: support SSE polling and server disconnects (SEP-1699)
This CL implements support for SEP-1699, adding the following new API:

- Event.Retry corresponds to the 'retry' field of a server sent event.
  This is used to signal client retry behavior.
- RequestExtra.CloseStream is set by the streamable transport to allow
  closing a stream.

The streamable server transport is updated to set CloseStream for calls,
allowing server operations or middleware to close the underlying
stream. If the configured 'reconnectAfter' delay is set, a 'retry'
message is sent prior to terminating the stream.

The streamable client transport is updated to read the configured delay.

Fixes #630