mcp: introduce options for AddTool
This change introduces a non-breaking API for passing
options to mcp.AddTool(). The first option introduced
is the jsonschema.ForOptions, which allows jsonschema
to use a specified mapping of Go types to static types
defined by the caller, which are used when generating
schemas instead inference via reflection. This is very
useful for example for Go structs implementing the
MarshalJSON and UnmarshalJSON interfaces in order to
represent the struct as a string. For example, imagine
a MonthID struct containing Year and Month as integers,
and whose serialization format implemented by MarshalJSON
is YYYYMM. From the jsonschema perspective the type
must be string, but the representation in Go can be
rich through a struct.

Fixes #658

Signed-off-by: Matheus Pimenta <matheuscscp@gmail.com>