mcp: don't break the streamable client connection for transient errors
When POST requests in the streamableClientConn return a transient error,
return this error to the caller rather than permanently breaking the
connection.

This is achieved by using the special sentinel ErrRejected error to the
jsonrpc2 layer. In doing so, the change revealed a pre-existing bug:
ErrRejected had the same code as ErrConnectionClosing, and
jsonrpc2.WireError implements errors.Is, so the two sentinel values
could be conflated. This is fixed by using a new internal code.

The new test required some additional machinery in our fake server: the
ability to handle multiple requests to the same logical key.

There's more to do for #683: we should also retry transient errors in
handleSSE.

For #683